--- a/src/client/opamListCommand.ml
+++ b/src/client/opamListCommand.ml
@@ -374,8 +374,8 @@ let rec filter ~base st = function
   | Block b -> filter ~base st b
   | And (a, b) ->
     let base = filter ~base st a in
-    base %% filter ~base st b
-  | Or (a, b) -> base %% (filter ~base st a ++ filter ~base st b)
+    filter ~base st b
+  | Or (a, b) -> filter ~base st a ++ filter ~base st b
 
 type output_format =
   | Name