With master (https://github.com/ocaml/opam/commit/f39739b9325ea10d5c733b14c8b14f6dcf46abf2):
```
$ time opam list --available --installable -s --all-version dune.2.9.1
dune.2.9.1
opam list --available --installable -s --all-version dune.2.9.1  54.60s user 0.60s system 99% cpu 55.278 total
```
With this PR:
```
$ time opam list --available --installable -s --all-version dune.2.9.1
dune.2.9.1
./opam list --available --installable -s --all-version dune.2.9.1  2.91s user 0.37s system 98% cpu 3.317 total
```

Incidently, this also makes `OpamListCommand.filter` correct. Previously it would simply ignore its `~base` parameter if the request did not start with a `And` case