@kit-ty-kate 2022-01-16
For `--conflicts-with` we get a 3x improvement. E.g. with opam master (https://github.com/ocaml/opam/commit/f39739b9325ea10d5c733b14c8b14f6dcf46abf2):
```
$ time opam list --installed --conflicts-with base-domains.base -s --all-version
js_of_ocaml-compiler.3.11.0
opam list --installed --conflicts-with base-domains.base -s --all-version  0.97s user 0.33s system 97% cpu 1.339 total
```
With this PR:
```
$ time ./opam list --installed --conflicts-with base-domains.base -s --all-version
js_of_ocaml-compiler.3.11.0
./opam list --installed --conflicts-with base-domains.base -s --all-version  0.39s user 0.05s system 96% cpu 0.461 total
```